requirements-lite.txt
torch>=2.1.0
numpy>=1.24.0
PyYAML>=6.0
imageio>=2.33.0
imageio-ffmpeg>=0.4.9
pillow>=10.0.0
tqdm>=4.66.0
